fix: fix like problems and make repository for dashboard summary
