+++ router/rest/auth.route.js
@@ -9,5 +9,6 @@ router.post('/register', AuthController.register)
 router.post('/login', AuthController.login)
 router.get('/me', authMiddleware, AuthController.getProfile)
 router.put('/me', authMiddleware, upload.single('photo'), AuthController.updateProfile)
+router.get('/dashboard', authMiddleware, AuthController.getDashboardSummary)
 
 module.exports = router
